Fix non-deterministic compiler seg faults when using parent classes with type variables and/or params only.
When a base class only has type variables and/or params, the function
resolution clean-up phase was removing such classes from the AST
because their constructors were not resolved (used).  This resulted in
accessing freed memory (for the class symbol and type) later in
compilation when subclasses checked the type of their super.  Since
there now appears to be some non-determinism in our compiler, we would
periodically see such tests succeed depending on whether or not that
piece of memory was being reused.

Fixed this by checking to make sure we didn't have such a class before
removing the symbol and type.  This is a little conservative as there
maybe be no subclasses or the subclasses may also be unused, but
checking for those cases would require more than a single pass over
all the types and doesn't seem worth it for this special case (tho
what's another pass when we're already doing like 10? :).

Removed the corresponding futures and updated a good file.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19422 3a8e244f-b0f2-452b-bcba-4c88e055c3ca