Correct the type of the non-blocking GET token.
Both chpl_comm_test_get_nb() and chpl_comm_wait_get_nb() had the type of
the GET token wrong.  What is passed to them is a value of the token
type, not a pointer to such a value.  This caused segfaults when the
code tried to dereference the parameter.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19430 3a8e244f-b0f2-452b-bcba-4c88e055c3ca