Fix for baseline test failures for iterators with INOUT arguments.
This is an error in the code which creates the advance() portion of an iterator
definition, which was masked by the --inline-iterators optimization.  When that
optimization is turned off, the wrong code is generated to replace a write 
to INOUT, OUT or REF args with a corresponding temporary.

Normally, the MOVE which originally writes into the arg is modified to write
into a temporary.  A statement is added, to copy the value in the temporary into
the field in the iterator record which represents the arg.  When the argument is
passed by reference, we need to load the address of the arg (from the field in
the iterator record) into a temporary and then write into the contents of that
temporary.

Code was added to identify and handle this case.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19451 3a8e244f-b0f2-452b-bcba-4c88e055c3ca