Added dsiLocalSlice() for Dimensional arrays.
This will make it support localSlice on Dimensional arrays
(but not domains).

This includes support for local slicing in block-cyclic 1-d descriptor.
It is implemented only when the slicing domain is not stridable.

Replicated and block 1-d descriptors did not need any modifications.

TODO 1: implement the above for stridable domains.

TODO 2: test for Dimensional with block 1-d descriptor(s).

TODO 3: Nail down and/or document the following semantic issue:
who ensures that the requested slice indeed is on the current locale?
In other words, need a check for that condition (perhaps disabled
upon --fast et al.).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19452 3a8e244f-b0f2-452b-bcba-4c88e055c3ca