Fix for valgrind regression exposed by standalone module initialization.
The field ModuleSymbol::initFn is essentially a cached pointer to the module
initialization function.  The declaration for that function is actually defined
in ModuleSymbol::block.  If the function is removed from the tree (using
Expr::remove()), the initFn field is not updated and becomes a dangling pointer.

I added the function cleanModulesList() (called by cleanAST()), to look for
module initialization functions which have been removed from the tree but not
yet deleted.  The initFn pointers are set to zero; in the suffix of cleanAST(),
the removed functions are then found and deleted.

I added a clause to Symbol::verify() in symbol.cpp to issue a fatal error if any
initFn pointer is left as a dangling reference at the end of a pass.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19453 3a8e244f-b0f2-452b-bcba-4c88e055c3ca