Adding -std=c99 to the RUNTIME_CFLAGS variable
This is required to use restrict in the runtime; it also matches what
we do for normal gnu compiles and arguably should have been doing all
along.

This fixes the build of the runtime for PrgEnv-gnu, but a warning at
link time remains.

Sorry that I didn't catch this in my testing -- I tested most of the
XT/XE compilers outside of the PrgEnvs for simplicity (all the modules
can be loaded simultaneously) and didn't anticipate that we'd need
flag changes to make it work within the PrgEnv wrappers.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19467 3a8e244f-b0f2-452b-bcba-4c88e055c3ca