Commenting out calls to getaddrinfo()
This routine was included as a forward-looking measure to support QIO
channels that target sockets, but is currently unused.  It results in
warnings for the PrgEnv-gnu compilations by default because static
linking is the default but getaddrinfo() requires a dynamic library to
be loaded.  So commented it out for the time being as a short-term
workaround.

I noted in comments: longer-term, we could move this routine into its
own .c/.o file so that the warning would only show up for users who
used sockets in their program and may care about the warning.
Otherwise, we could look into squashing it.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19468 3a8e244f-b0f2-452b-bcba-4c88e055c3ca