Fixing uninitialized var issue that was causing --baseline failures
--baseline testing was having problems with the new I/O system because
of a read of an uninitialized 'showpointzero' value in the style
record.  It seems that this is due to the fact that the field was not
declared in the extern record iostyle declaration, presumably
indicating a problem due to a compiler-introduced copy (that was
perhaps being optimized away?).  In any case, this change simply
declares the field which also makes the problem go away, seemingly.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19471 3a8e244f-b0f2-452b-bcba-4c88e055c3ca