Really turn off tcmalloc messages for large allocations.
This is a followup to r19432.

When I disabled large allocation reporting in 19432, I missed the fact that
the logic imposes a minimum of kPageSize for the threshold size above which
allocations are reported.  But the code which compares the threshold to this
minimum doesn't handle the case of a zero threshold, which is supposed to turn
off the reporting altogether.  As a result we were still getting reports of
large allocations, but with a smaller threshold than before.  Fix the bug.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19488 3a8e244f-b0f2-452b-bcba-4c88e055c3ca