Fixes util/config scripts
Continuing r19463, I've put the new targets
in runtime/etc/Makefile.include to get it to
print out the compile or link line. Also
updated the path that the compileline script
checks for detecting configuration errors
(ie, it complains when ../.. != $CHPL_HOME).

I still need to know if ./util/config/makeSysBasicTypes
produces the correct output on all platforms...
If so, we can proceed with using that script to
generate a .chpl file that is always included.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19490 3a8e244f-b0f2-452b-bcba-4c88e055c3ca