Split SSCA2 testing into 5 distinct tests (one per graph type).
This change uses compopts to compile only a single graph type at a
time for each of the 5 graph types (RMAT, 1D-4D torus).

- Modified the config params and SSCA2_mainc.chpl to enable this.

- Added new .good files for each graph type (since we have different
  debug output for each).

- Changed the Makefile a little to enable easier adding of other
  compile-time flags.

- Added a comment in the prediff script regarding the testing change.

- Updated the README

The performance test is not affected, as we are only compiling RMAT
graph for performance testing.

This should help with the periodic timeouts for SSCA2_main.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19521 3a8e244f-b0f2-452b-bcba-4c88e055c3ca