Making Nanos++ tasking yield on active messages
While sending a mail to the MassiveThreads group this week I was
grepping the sources and realizing that we'd never defined the hook in
the Nanos tasking layer to yield tasks when active messages are fired
off.  This is the reason that simple programs like some of the
distributed memory hello worlds have been deadlocking when numLocales
is greater than numCores -- the worker threads would get blocked in
active message calls.  Whoopsie.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19522 3a8e244f-b0f2-452b-bcba-4c88e055c3ca