Changed sgn() in Math.chpl to return u/int(8).
The motivation is that the result of sgn() fits in uint(8)
for a uint(?) argument and in int(8) for any integer or real argument.
u/int(8) can then be implicitly converted to most other types of interest.

uint(8) wouldn't be implicitly converted to int(8), but it feels a small
price to pay for knowing that the result of sgn(uint(?)) is non-negative.

If sgn() returns an int(32 or 64) or real(?), an explicit cast would be
required to use it in some contexts, e.g.

var a, b: int(8);
var c = a..b by sgn(....);

Why is the result's bitwidth 8? A: that's the smallest bitwidth
that's currently supported. If uint(1) and int(2) were available,
we would use those instead.

We also considered having sgn() return an enum {NEG=-1, ZERO=0, POS=1},
but chose against it - see below.


Here are snippets from an email discussion on this subject, just
for a record.

-----------
For a function that maps into a limited range, I think it makes perfect sense to have the return type of the function reflect that limited range.

In other words

Given y = sgn(x) as a mapping from reals or integers (of any size) into {-1, 0, 1}, either create a type to represent that limited set, or choose the smallest type that can represent the result -- in this case, int(8).

Supposing there was an implicit cast from enumerated types to integer types, having the result lie in 

enum sgn_result {
  NEG = -1,
  ZERO = 0,
  POS = 1
}

would be excruciatingly correct.

-----------
I like this enum proposal as well.  Since our enums are essentially 
param-like and since params are more coercion-friendly than 
variables/constants, this seems like it could fall out pretty easily. I 
also like that it doesn't embed "8" into the signature, which feels pretty 
arbitrary (particularly if we end up supporting int(4)/int(2) someday...)

-----------
Returning an enum sounds fine by me.

However, there is a glitch whereby the result of sgn() of the envisioned
enum {-1, 0, 1} would not be assignable to an unsigned int. The spec
says: "An expression of an enumerated type can be implicitly converted
to an integral type, provided that all of the constants defined by the
enumerated type are representable by the integral type."

Because of that, having sgn(uint(?)) return some uint might be preferable.

-----------
Oh, that's a good point.  Yeah, this pushes me back towards the 
int(8)/uint(8) proposal.

Or toward special-casing the uint case to cast the enum to a uint(8) 
before returning it (downside: more asymmetric).  Or creating a second 
enum for unsigned signs that just has the 0, 1 cases (downside: more 
enums).

I think if I had to make a snap decision, I'd go with the int(8)/uint(8) 
proposal for simplicity/consistency, though I'm not opposed to the other 
two approaches either.

-----------


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19530 3a8e244f-b0f2-452b-bcba-4c88e055c3ca