Introduced a "wrapper" distribution (WrapperDist) that will redirect most operations to its wrappee; touches on the existing Dimensional.
This is a checkpoint of work in progress.

* WrapperDist, WrapperRectDom, WrapperArr classes are defined in n.chpl.

* Dimensional seems to use WrapperDist successfully to implement reindexing.
Support for rank change in WrapperDist is under development.

* Disabled the assertion (myPiece.length > 0) in DimensionalDom leader iterator
because it was tripped over in slicing/test_strided_slice1.chpl
from the robustness suite. We could check whether there is a way
to enable it under some conditions.

* Introduced const index0 = indexx - bbStart in dsiIndexToLocale1d()
for 1-d block descriptor to avoid overflows in the robustness suite.

* Added dsiCreateReindexDist1d() for 1-d replicated and block
descriptors. Those were originally used in genuine support
for reindexing in Dimensional. However, that did not apply
to 1-d block-cyclic and so, for the moment at least,
they are not used - they are superceded by WrapperDist.

Design notes - WrapperDist.

* One motivation in creating this distribution is proof of concept:
Given a domain map class, can one of its DSI methods
(e.g. dsiCreateReindexDist) return an object coming from
another domain map?

* The functionality in WrapperDist that can be shared between
wrapper-based support for reindexing and rank change is:
- mostly, privatization
- also whole domain-related operations.

* Ideally this would be an inheritance hierarchy, with an abstract
superclass defining this shared functionality and subclasses giving
the specifics of reindexing, rank change, etc.
  For no good reason other than I thought about it first, I instead
wrote a single class. Its purpose (reindexing/rank change/...)
is defined by the 'kind' param.
  The upside is that this bypasses the performance concerns
that we subconciously relate to Chapel's implementation of
inheritance/dynamic dispatch. The downside is that coding
is a little more complex.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19536 3a8e244f-b0f2-452b-bcba-4c88e055c3ca