Condense compopts/execopts info in failed test error messages.
At Brad's request, rather than printing out all the compopts and
execopts when a test fails, I've changed it to simply encode the
variation that failed.

For example,

   [Error matching program output for tmp/blah (compopts: 1, execopts: 2)]

means that the error happened with the compiler flags given on the
first line of blah.compopts while running it with the flags given on
the second line of blah.execopts.

If you see a 0 for one of those numbers, that means that there was no
.compopts or .execopts file.  You'll never have 0 for both (it's
omitted, since that is the default case).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19544 3a8e244f-b0f2-452b-bcba-4c88e055c3ca