Further implementation of WrapperDist; ensured it passes the robustness suite.
The robustness suite: needed one mod to hpl.chpl (details below).

TODO items are listed in d.chpl, n.chpl.

Mods by file:

  d.chpl (DimensionalDist)
     added a couple of dsi methods

  n.chpl (WrapperDist)
     implemented more functionality + fixes
     made the code tidier
     left debugging code in, for now

  f.chpl (1-d block-cyclic descriptor)
     suppressed a compiler warning about inefficiency,
     so it passes the robustness test suite

  COMPOPTS
     enables that warning for tests in this directory

-----------

The mod to hpl.chpl that's needed for it to work with WrapperDist
- not checked in, but used for robustness testing -
is this:

-proc LUFactorize(n: indexType, Ab: [1..n, 1..n+1] elemType,
+proc LUFactorize(n: indexType, Ab,

Specifying the range for Ab inserts implicit reindexing from the actual
to the formal. Since the actual is mapped with Dimensional, this
results in a Wrapper-mapped array. With my change, it continues to be
a Dimensional-mapped array.

That array is then passed to panelSolve, which expects an array
mapped with Dimensional:

 proc panelSolve(Ab: [] ?t,
                panel: domain(2, indexType) dmapped Dist2D,
                piv: [] indexType) {

If our array is Wrapper-mapped, a compiler error occurs.
The same issue is exposed in updateBlockRow.

Since examples/hpcc/hpl.chpl does not impose this requirement
(and furthermore LUFactorize accepts any array),
I worked around it.

I also checked that Dimensional works when LUFactorize
is left unchanged, but instead panelSolve/updateBlockRow
modified to have the panel/tl/tr args be of the type 'domain',
like they are in examples/hpcc/hpl.chpl.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19548 3a8e244f-b0f2-452b-bcba-4c88e055c3ca