Change capture tests to again illustrate the issue described in the future.
After the QIO commit (r19458), the future capture.chpl started passing
because the argument to the work() function was no longer passed by
reference.  Michael explained that this was probably because writeln()
now takes it's arguments by value.

To illustrate the point of the future, I managed to get the compiler
to pass the argument by reference by adding a barrier via sync and
single vars.  I actually don't understand why this works, but I have
not investigate too deeply.  The heap allocation code has not change
much since the initial commit by Steve, and there are not too many
clues.

Updated to future to note this weirdness.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19555 3a8e244f-b0f2-452b-bcba-4c88e055c3ca