Added 'cpl.c', a C version of HPL. The HPL version is taken from release/examples/hpcc and considered as if compiled with --serial.
I'd note two deviations of the C version:

* The two reductions are almost completely inlined, except the function
  that computes the value for each index is passed as an argument.

* Domains and ranges are passed around by value throughout.
  Arrays are passed by reference, however.

The C compilation command is:

   gcc -std=c99 [optimization options] cpl.c

The Chapel compilation command is:

   chpl -M [path to a dir where HPCCProblemSize.chpl is located] \
     [stdlib.h may need to be specified] \
     [optimization options] --serial cpl.c

'hpl.chpl' is a modification of a copy from release/examples/hpcc.
The modifications have been done to ensure that the Chapel and C
versions perform the same computation.

It is marked with .notest because currently the output includes the
timing and so is nondeterministic.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19563 3a8e244f-b0f2-452b-bcba-4c88e055c3ca