Rewrite of fiterSuppressions script.
With the new compopts/execopts summary reporting, this script was
missing suppressions for tests that had previously been summarized
without compopts or execopts.

While looking at this issue, I had a difficult time leaving the script
as is because it was so horribly inefficient (sorry Brad), so I
rewrote it completely in Python (again, sorry Brad).

With this change, you can now also specify a particular
compopts/execopts combo for a given test in the suppressions file.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19575 3a8e244f-b0f2-452b-bcba-4c88e055c3ca