Removed debugWrite* capabilities
Based on consensus across the team and the fact that they don't fit
the new model for I/O particularly well, I've removed the debugWrite*
capabilities that we have supported in the I/O modules for some time.
These were essentially glorified printf+fflush calls that were
designed to support lightweight local printing without relying on sync
vars and active messages.  But they also have the downside of only
supporting a subset of Chapel types.  Now that our extern capability
is better, printf() and fflush() can be used directly if someone
wanted this capability.  E.g.:

     extern proc printf(x...);
     extern proc fflush(x);
     extern var stdout: opaque;

     printf(...);
     fflush(stdout);

Or, using the new QIO capability, Michael suggests that one could
"always open up a (possibly unlocked) channel (something like this):

 var mystdout = openfd(1).writer();
 mystdout.write("something");
 mystdout.flush();

The main work in this change was updating tests.  I took the stance of
(a) using printf("%s\n", ...) in cases where that worked (with no
fflush() because I didn't seem to need them), (b) using writeln() in
cases where multiple locales were printing and the results seemed
race-y if I didn't.  With more effort, we could probably convert more
of the printf() cases to write()s, but I didn't have time to put in
that effort this time around and hoped that the original choice of
debugWrite() was made for a good reason...




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19581 3a8e244f-b0f2-452b-bcba-4c88e055c3ca