Update gasnet .good files for RADOpt tests.
These tests started failing for gasnet after the QIO commit in a good
way.  There is one fewer remote get reported for all remote array
accesses.  Before the QIO commit, the privatization ID was being
passed by reference to my test function because it was also passed to
writeln() which used to take it's arguments by reference.  Now
writeln() takes its arguments by value, so the (wide reference to the)
pid is passed by value and no communication is needed to get the pid
value.  

This is the same issue encountered with my capture tests (r19555).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19584 3a8e244f-b0f2-452b-bcba-4c88e055c3ca