Implement a more portable way to query the number of cores.
Added numCores.prediff, a python script that uses the multiprocessing
module to get the number of cores on the machine.  The script writes
the output to numCores.good, so that in cases where it doesn't work or
is wrong (e.g., cross compilation), we can add a platform-specific
good file.

All the .good files have been removed, and the new CLEANFILES will
remove numCores.good when cleaning.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19588 3a8e244f-b0f2-452b-bcba-4c88e055c3ca