Updated nightly test script to 'make' in modules directory
My commit yesterday broke the nightly testing script because it
performs explicit 'make' steps on the compiler and runtime directories
which means that the new 'modules' make step was never invoked.  I
fixed it by adding an explicit make step for the modules since we had
compelling reasons for explicitly separating the compiler and runtime
build steps (such as supporting a mode in which we simply make sure we
can build the compiler with a given compiler).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19591 3a8e244f-b0f2-452b-bcba-4c88e055c3ca