More makefile clean-up.
The bulk of this commit replaces CHAPEL_ROOT with CHPL_MAKE_HOME
making it more consistent with other variable naming (plus Brad didn't
like the vowels :)

I also implemented a policy for setting CHPL_MAKE_HOME, as it was
previously inconsistent and/or redundant.  Any file named 'Makefile'
is assumed to be one that can be used directly (rather than by
inclusion).  If CHPL_MAKE_HOME is not defined, then the Makefile
defines and exports it relative to it's current (absolute) path.  The
top level Makefile will always set CHPL_MAKE_HOME.

This has the addedbenefit of making it possible to use a sym link to a
no-release third-party directory rather than having to make a copy of
it, because 'make X' in $CHPL_HOME/third-party sets CHPL_MAKE_HOME.
We might want to eventually replace the explicit copy of mpich1 in our
nightly testing to take advantage of this.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19595 3a8e244f-b0f2-452b-bcba-4c88e055c3ca