Submitting reduction bug submitted by Mark Potts and summer interns
In the ongoing attempt to clean up my inbox, I'm filing away a bug
that we never looked at this summer.  It seems that when overloading
the ^ operator for a user-defined type we get an internal error when
trying to apply a ^ reduction to an array of said type.  Strangely,
this does not happen when using the + operator.

I started trying to diagnose this and got as far as finding that we
were trying to call a min() primitive on the user's type.  I'm not
sure why we would be trying to do that (possibly to compute the
identity value?  But strange for a ^ operator...).  But in any case,
this should not be an internal error.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19597 3a8e244f-b0f2-452b-bcba-4c88e055c3ca