Appling a qthreads patch from Dylan designed to make it work with QIO
This patch permits Qthreads to work cleanly with the new QIO code.
The breaking point was that standard modules (like IO) are currently
initialized outside of any Chapel tasks by a program's main thread
(i.e., before the tasking layer is requested to create a task to run
chpl_main()).  But the initialization code currently results in calls
to chpl_task_getId() which was not well-defined for Qthreads (and for
that matter most tasking layers) outside of a task.  This patch
introduces a sentinel value for that purpose.

FWIW, we are also considering a change in which the initialization of
the standard modules is done within the main task, though there are
other ordering issues to work through before that works.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19606 3a8e244f-b0f2-452b-bcba-4c88e055c3ca