Made 'def' no longer a keyword. Removed the corresponding computation of whether a function is a proc or iter.
Basically I followed the instructions in r18219 for
"Compiler modifications to be done to move out of the "transition" period".

Removed those tests in test/functions/vass/proc-iter/
that were specific to the transition period.
The remaining tests do some basic checking of proc/iter,
mostly that the compiler generates the correct errors
for violations of semantic rules.

'ProcIter' in include/yy.h and parser/chapel.ypp can easily be just 'bool',
since now it merely distinguishes procs from iters.
So far it's been an enum, and I did not feel it was important to change that.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19608 3a8e244f-b0f2-452b-bcba-4c88e055c3ca