Replaced BaseAST::lineno with BaseAST::astloc, presently a struct with a single 'lineno' member.
This is a first step towards having each AST node store its own filename.
Today, a node's filename is always computed as that of its containing module.
This does not work for nodes that resulted from inlining, which we aim to fix.

The changes in this commit are as follows:

* Introduced
    struct astlocT
as the type that will store a node's location in the source code.

It seems important that 'astlocT' retain the same copying semantics
as integers, because we will be assigning it and passing as arguments
as we did BaseAST::lineno.

* In BaseAST, replaced 'int lineno' with 'astlocT astloc'.

* Replaced all line-number queries and most filename queries to AST nodes with
    BaseAST::linenum()
    BaseAST::fname()

(In some cases it is not clear whether the code's author wanted
a node's filename or its module name or module filename. I left
those cases unchanged.)

* 'int currentLineno' --> 'astlocT currentAstLoc'.
SET_LINENO() updates the entire astlocT (whatever it may be).

(I left the name unchanged to reduce the diff footprint and also
because we might want in the future to change how we propagate
locations. We can rename it at that time.)

* Replaced
    void reset_line_info(BaseAST* baseAST, int lineno);
with
    void reset_ast_loc(BaseAST* destNode, astlocT astloc);
    void reset_ast_loc(BaseAST* destNode, BaseAST* sourceNode);

I had to modify an argument (int -> astlocT), so I chose to rename
the function as well.

* Renamed a local variable 'fname' to 'flname' so it stays away
from potential grep queries.


In this commit, the functionality has not changed: astlocT stores
only the line number. Indeed,

    struct astlocT     == { int lineno; }
    BaseAST::linenum() == astloc.lineno
    BaseAST::fname()   == getModule()->filename

To achieve our goal of having each AST node store its own filename,
we will have astlocT represent the filename as well.
This could be done directly by storing a pointer or via a space-saving
representation e.g. a table index.


I added comments tagged with "astloc" to mark some spots that
will need to be revisitied when each AST node stores its own filename.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19612 3a8e244f-b0f2-452b-bcba-4c88e055c3ca