Have paratest.server detect a fatal() in paratest.client.
Before this change, if an instance of paratest.client failed,
i.e. exited via its fatal(), it did not create a syncfile.
Since paratest.client relies on syncfiles to know which clients
are still working, it would wait forever for the missing syncfile.

This change has paratest.client create an alternative syncfile
whose name indicates it's created upon a failure.
paratest.server detects that syncfile and takes that into account:
- reduces by 1 the total number of clients,
- does not spawn any more jobs on that client.

We could choose an alternative behavior, whereby the failed client
creates a regular syncfile. Then paratest.server handles it as a
finished job like any other, and so continues spawning more jobs there.

While making this change, I shuffled a couple of variables in the scripts
and removed a seemingly-obsolete comment.

Note regarding 'wait' in free_workers() in paratest.server:
In my testing, it would occasionally indeed wait for something
(after a failed paratest.client), preventing the famous dots
from printing in the shell window.
I do not claim any understanding, but it does not seem like
a big issue. FWIW, according to the comments in the code and in r13881,
it "cleans up zombied child processes." It's "in a loop over the
files that represent children who have finished their work and exited."
I believe I observed the no-dots-printing behavior even when running
the scripts without the change in this commit.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19618 3a8e244f-b0f2-452b-bcba-4c88e055c3ca