Adding -Wno-char-subscripts flag to gcc developer warnings
With the QIO change, we started tripping over warnings using -Wall on
cygwin (only), probably due to gcc version number, related to indexing
into arrays using characters.  Michael preferred squashing the warning
rather than casting all the chars to be ints and I didn't object, so
this flag does so.  This should only affect developers, not end-users.
If anyone has portability problems with it (like versions of gcc that
you're using that don't support it), please give a shout.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19620 3a8e244f-b0f2-452b-bcba-4c88e055c3ca