Added the --preserve-inlined-line-numbers flag.
When this flag is given, the compiler will preserve the file name
and line number in the nodes of a subtree that's being inlined.
By default (this flag is off), the behavior prior to this change
is preserved. Namely, the file name/line number of all nodes
of a subtree being inlined get replaced with the file name/line number
of the caller expression.

Implementation note: all the flag does is control two calls to
reset_ast_loc(), formerly reset_line_info() - one in lowerIterators.cpp,
one in inlineFunctions.cpp.

Why is this flag not enabled by default? Turns out in some cases
the behavior without this change is preferred over the behavior
with this flag on. Here is an example:

Say there is this call at f.chpl:23:

  A.this(i)

(This could be indexing into the array A with index i, but anyway.)

Background: The callee gets inlined. The callee calls dsiAccess(), and
that call gets inlined too.
  dsiAccess invokes halt("array index out of bounds: ", ...), which,
after inlining, becomes  __primitive("chpl_error", ...).
  That primitive call is augmented in the insertLineNumbers pass
to pass its caller's file name and line number.

The behavior without this change is that the very last caller's
file name and line number are those of the call that triggered
all this inlining. That is, A.this(i).
So the user gets the nice error indicating the location of A.this(i):
"array index out of bounds at f.chpl:23".

The behavior with this flag ON is that the error message indicates
the location of the 'halt' call, which is DefaultRectangular.chpl:552.
This conveys little useful information to the user.

This scenario can be observed in several tests in the test system.
Therefore I am leaving this change off by default.
(Which means that by default the file name pointer in each AST node
is really not needed. But it is needed when the flag is turned on.)

The behavior with the flag off simply happens to work well in that
test example. That's because the location printed by halt(),
chpl_error() et al. is determined by inlining, rather than
any language-based rules.

Ideally we'd like to see the location of the innermost call that's
still in the user code. But this is not guaranteed even with the
flag off.  E.g. if f.chpl:23 has the call bar() and bar() is defined,
say, at DefaultRectangular.chpl:678, then the "array index out of bounds"
error message will display DefaultRectangular.chpl:678 (an internal
module), rather than the user call at f.chpl:23.

We haven't come up with a way to do it right, therefore
the --preserve-inlined-line-numbers flag is offered as a quick workaround.
With the flag ON, the location displayed in an error message will be
less useful, but each stack frame of the stack trace in the debugger
will supposedly have a more useful location.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19621 3a8e244f-b0f2-452b-bcba-4c88e055c3ca