Removing _POSIX_SOURCE macro for XMT compilation
[concept reviewed by Michael and Sung]

I'm pulling a #define of _POSIX_SOURCE that came in with the QIO patch
because it breaks compilation on an XMT and seems to be a deprecated
macro.  If it breaks other things we can deal with those as we come
to it, though we don't anticipate it will.

Unfortunately, the XMT build will need more work before it's working
again (Michael's on it), but this seemed worthwhile to commit as a
first step.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19623 3a8e244f-b0f2-452b-bcba-4c88e055c3ca