Added a bug future whereby a default class type constructor does not get resolved, under a certain pattern of 'use' declarations.
SSCA2-with-Private-1.chpl is a slimmed down example of what we experienced
when trying to use PrivateDist in SSCA2_kernels.

The example is simple:

    module SSCA2_main {
      proc main () {
        use SSCA2_kernels;
      }
    }
    module SSCA2_kernels { 
      use PrivateDist;
    }

but produces

    <the location varies>: error: unresolved type specifier 'Private()'

The actual code checked in has a couple of more lines in it,
just to make sure things work when they do.

Putting two modules in a single file seems to work, whereby the compiler
takes the first one to be the main module (because it has main()),
and the second one is available because it has already been parsed.

SSCA2-with-Private-2.chpl illustrates the same issue, but
has the tree of the 'use' statements that is (intended to be) directly
a subset of what's in test/release/examples/ssca2.
(Still not perfect because the 'use' tree in SSCA2 has more occurrences
of the modules listed in SSCA2-with-Private-2.chpl.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19625 3a8e244f-b0f2-452b-bcba-4c88e055c3ca