Created a trimmed copy of test/release/examples/ssca2 for graph I/O work for SSCA#2.
This commit is intended as the reference point against which to diff
the future graph I/O changes.

Only three files are copied from test/release/examples/ssca2:
  SSCA2_RMAT_graph_generator.chpl
  analyze_RMAT_graph_associative_array.chpl
  SSCA2_main.good --> graphio.good (trimmed)

the rest are used by reference or via wrappers or substituted
by empty declarations in graphio.chpl.
E.g. execute_SSCA2() is replaced with a version that just does a writeln().

I added a writeln() to each of the two copied .chpl files
to confirm that the copies are used, not the original versions
from test/release/examples.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19631 3a8e244f-b0f2-452b-bcba-4c88e055c3ca