Fix for GASNET+CHPL_DEVELOPER build problems.
Gasnet builds add -Winline to the warnings, which
causes Bad Things with CHPL_DEVELOPER-enabled -Werror
when an inlined function is too big. That's usually
useful information, but sometimes unavoidable.

Continuing r19641, memory allocation and freeing
routines are now inline, but GASNet builds were
reporting errors.

The workaround is to use ___always_inline
instead of inline, which currently works for gcc.
Nonetheless, I'd rather see a relaxation of
-Werror with e.g. -Wno-error=inline - which would
cause the warning to still be printed - but that
only works on new enough GCCs.

Patch requested by Sung.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19650 3a8e244f-b0f2-452b-bcba-4c88e055c3ca