Removing duplicate files from release tarball, along with a few developer files
At some point when the gen_release script was changed to include the
entire third-party directory (probably by me), we didn't remove the
explicit request to include certain files from the third-party
directory, which seems to have had the effect of adding them as
secondary symbolic link entries.  This seemed to cause problems
unpacking the release tarball for me on my Mac (Lion) this week, so
this commit removes the redundancy (which is neater anyway).

While here, I noticed that the same change was causing some developer
files (README.devel and Makefile.devel*) to be included in the release
which was also unintentional, so I added steps to remove them.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19652 3a8e244f-b0f2-452b-bcba-4c88e055c3ca