Removing an extraneous -I flag
Our current Makefile rules were compiling the generated code using a
flag -I-I. due to the specification of -I both within and outside of a
list of include paths.  This mostly didn't cause any problems (due to
a redundant -I. elsewhere in the path), except on PGI where it caused
some weird errors to come out of the compiler (the reason that PGI
testing has recently been failing with compile-time differences rather
than the execution-time differences we were seeing just previously.
To fix this, I removed the -I that lives outside the variable name to
keep things a little simpler.  I also took this opportunity to rename
CHAPEL_RT_INC_DIR and CHAPEL_RT_LIB_DIR to CHPL_RT_*... to get rid of
one more place in the code where the vowels are used.

Another cleanup that could probably be done is to remove the extra -I.
mentioned just above.  I didn't get that in before testing, and wanted
to clean up PGI before I got distracted again, but will take a look at
that next.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19661 3a8e244f-b0f2-452b-bcba-4c88e055c3ca