Remove an unused variable.
When I committed r19664 I hadn't yet been able to build the PVM-based
comm layer.  Once I managed to do that I discovered that I'd left behind
a definition for a variable that was no longer used.  Remove it, mostly
because it's good form, but also because if gcc is the target compiler
its "defined but not used" warning for this variable causes the build to
fail.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19675 3a8e244f-b0f2-452b-bcba-4c88e055c3ca